requirements.txt
absl-py==1.4.0
aiocache==0.11.1
aiofiles==0.8.0
aiohttp==3.8.4
aiohttp-socks==0.8.0
aiomysql==0.1.1
aiosignal==1.3.1
aiosqlite==0.17.0
alembic==1.11.1
amis-python==1.0.8.post2
anyio==3.7.0
APScheduler==3.10.1
arrow==1.2.3
asgiref==3.7.2
async-timeout==4.0.2
asyncio==3.4.3
asyncio-dgram==2.1.2
asyncpg==0.27.0
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bbcode==1.1.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
blinker==1.4
cachetools==5.3.1
cairocffi==1.6.0
CairoSVG==2.7.0
cashews==6.2.0
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.1.0
chevron==0.14.0
click==8.1.3
cloud-init==19.1
cmake==3.27.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.0
cookiecutter==2.1.1
cryptography==3.4.8
cssselect2==0.7.0
cycler==0.11.0
Cython==0.29.35
dbus-python==1.2.18
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
dnspython==2.2.1
ecdsa==0.18.0
emoji==2.8.0
exceptiongroup==1.1.1
fake-useragent==1.1.3
fastapi==0.101.1
filelock==3.12.2
fonttools==4.40.0
frozenlist==1.3.3
fuzzywuzzy==0.18.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.57.0
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httplib2==0.20.2
httptools==0.5.0
httpx==0.24.1
humanize==4.7.0
hyperframe==6.0.1
hyperlink==21.0.0
icalendar==5.0.7
icmplib==3.0.3
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
iso8601==1.1.0
jeepney==0.7.1
jieba==0.42.1
Jinja2==3.1.2
jinja2-time==0.2.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lit==16.0.6
loguru==0.6.0
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
mcstatus==10.0.3
MonthDelta==0.9.1
more-itertools==8.10.0
mpmath==1.3.0
ms-api==0.10.226.post2
msgpack==1.0.5
multidict==6.0.4
nb-cli==1.1.2
netifaces==0.11.0
networkx==3.1
nonebot-adapter-kaiheila==0.2.6
nonebot-adapter-onebot==2.2.1
nonebot-plugin-anime-trace==0.1.1
nonebot-plugin-animeres==1.0.3
nonebot-plugin-answersbook==0.2.0
nonebot-plugin-antiinsult==0.3.3
nonebot-plugin-apex-api-query==23.6.23
nonebot-plugin-apexranklookup==0.1.6
nonebot-plugin-apscheduler==0.2.0
nonebot-plugin-AutoRepeater==0.1.4
nonebot-plugin-bfinfo==0.0.5
nonebot-plugin-bilibilibot==2.3.2
nonebot-plugin-blacklist==0.3.5
nonebot-plugin-bottle==1.0.1.3
nonebot-plugin-cartoon==0.1.0
nonebot-plugin-chatrecorder==0.2.4
nonebot-plugin-cloudsignx==2.0.0
nonebot-plugin-CyberSensoji==0.0.2
nonebot-plugin-datastore==1.0.0
nonebot-plugin-dialectlist==1.4.0
nonebot-plugin-dog==0.2.7.1
nonebot-plugin-error-alert==0.1.2
nonebot-plugin-gocqhttp==0.6.11
nonebot-plugin-gocqhttp-cross-machine-upload-file==0.1.5
nonebot-plugin-groupmate-waifu==1.2.6
nonebot-plugin-guild-patch==0.2.3
nonebot-plugin-htmlrender==0.2.1
nonebot-plugin-HttpCat==0.0.2
nonebot-plugin-Imagelabels==1.0.0
nonebot-plugin-imageutils==0.1.17
nonebot-plugin-impact==0.0.10
nonebot-plugin-jrrp2==1.6
nonebot-plugin-learning-chat==0.2.3
nonebot-plugin-localstore==0.4.1
nonebot-plugin-lolmatch==0.3.0.2
nonebot-plugin-majsoul==0.2.1.post1
nonebot-plugin-megumin==0.0.5
nonebot-plugin-miao==1.1.0
nonebot-plugin-picmcstat==0.3.5
nonebot-plugin-picsearcher==0.1.10
nonebot-plugin-picstatus==0.5.1
nonebot-plugin-r6s==0.4.2
nonebot-plugin-random-stereotypes==0.1.0
nonebot-plugin-reboot==0.1.4
nonebot-plugin-reborn==0.4
nonebot-plugin-russian-ban==1.0.6
nonebot-plugin-searchbiliinfo==1.7.9
nonebot-plugin-send-anything-anywhere==0.2.7
nonebot-plugin-setu-collection==0.1.4
nonebot-plugin-shindan==0.4.0
nonebot-plugin-simplemusic==0.1.7
nonebot-plugin-status==0.7.1
nonebot-plugin-tarot==0.4.0.post4
nonebot-plugin-tortoise-orm==0.0.1a3
nonebot-plugin-txt2img==0.2.2
nonebot-plugin-what2eat==0.3.6
nonebot-plugin-whateat-pic==1.1.6
nonebot-plugin-xingzuo==1.0.3
nonebot-plugin-xiuxian-2==2.9.2.2
nonebot-plugin-ygo==1.0.2
nonebot2==2.0.1
noneprompt==0.1.9
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.0
opencv-python==4.8.0.76
opencv-python-headless==4.7.0.72
packaging==23.1
pandas==2.0.3
pexpect==4.8.0
pil-utils==0.1.7
Pillow==9.4.0
platformdirs==2.6.2
playwright==1.35.0
prompt-toolkit==3.0.38
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycocotools==2.0.2
pycparser==2.21
pydantic==1.10.9
pyee==9.0.4
pyfiglet==0.8.post1
pygal==3.0.0
Pygments==2.15.1
PyGObject==3.42.1
pygtrie==2.5.0
PyHamcrest==2.0.2
PyJWT==2.3.0
pymdown-extensions==10.0.1
PyMySQL==1.1.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pypika-tortoise==0.1.6
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==1.0.0
python-jose==3.3.0
python-magic==0.4.24
python-markdown-math==0.8
python-slugify==8.0.1
python-socks==2.3.0
pytz==2022.1
PyYAML==5.4.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scipy==1.11.2
seaborn==0.12.2
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sos==4.4
soupsieve==2.4.1
SQLAlchemy==2.0.17
ssh-import-id==5.11
starlette==0.27.0
StrEnum==0.4.12
sympy==1.12
systemd-python==234
tensorboard==2.14.0
tensorboard-data-server==0.7.1
tensoul==0.1.3
text-unidecode==1.3
thop==0.1.1.post2209072238
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.8
torch==2.0.1
torchvision==0.15.2
tortoise-orm==0.19.3
tqdm==4.66.1
triton==2.0.0
Twisted==22.1.0
typing_extensions==4.7.0
tzdata==2023.3
tzlocal==5.0.1
ubuntu-advantage-tools==27.9
ubuntu-drivers-common==0.0.0
ufw==0.36.1
ujson==5.7.0
ultralytics==8.0.157
unattended-upgrades==0.1
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
virtualenv==20.17.1
wadllib==1.3.6
watchfiles==0.19.0
watchgod==0.8.2
wcwidth==0.2.6
webencodings==0.5.1
websockets==10.4
Werkzeug==2.3.7
wget==3.2
wrapt==1.15.0
xkit==0.0.0
yarl==1.9.2
youth-version-of-setu4==0.0.30
zipp==1.0.0
zope.interface==5.4.0
